Removed some compiler warnings.

+++ AbstractViewController.m
@@ -50,10 +50,10 @@
 - (void)updateWordCount
 {
 	NSArray *words = [textView.text componentsSeparatedByCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
-	int count = words.count;
+	NSUInteger count = words.count;
 	if (((NSString *)[words objectAtIndex:0]).length == 0) --count;
 	if (count > 0 && ((NSString *)[words lastObject]).length == 0) --count;
-	wordCountLabel.text = [NSString stringWithFormat:@"Word count: %d  (recommended 150-250)", count];
+	wordCountLabel.text = [NSString stringWithFormat:@"Word count: %ld  (recommended 150-250)", count];
 }
 
 @end
